requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
aiosqlite==0.19.0
